consumer/requirements.txt
kafka-python
requests
opensearch-py